yggdrasil_ai/requirements.txt
openai>=1.3.0
pinecone>=3.0.0
python-dotenv>=1.0.0
rich>=13.0.0
streamlit>=1.30.0
ratelimit>=2.2.1
